loan_pj/requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn  
scikit-learn
pymongo
from_root
dill
PyYAML
fastapi
python-multipart
uvicorn
imblearn
-e .